Corrected the json lint file and separated the pure dom functions from the index
